requirements.txt
anthropic>=0.40.0
moviepy>=1.0.3
Pillow>=10.0.0
numpy>=1.24.0
gradio>=4.0.0
requests>=2.31.0
PyJWT>=2.8.0
aiohttp>=3.9.0
aiofiles>=23.0.0
